refactor: seperate class file
