+++ packages/effects-core/src/animation/index.ts
@@ -1,1 +1,2 @@
 export * from './color-playable';
+export * from './animation-clip';
