+++ packages/effects-core/src/plugins/timeline/playable-assets/activation-playable-asset.ts
@@ -1,0 +1,12 @@
+import * as spec from '@galacean/effects-specification';
+import { effectsClass } from '../../../decorators';
+import type { Playable } from '../../cal/playable-graph';
+import { PlayableAsset } from '../../cal/playable-graph';
+import { ActivationPlayable } from '../playables';
+
+@effectsClass(spec.DataType.ActivationPlayableAsset)
+export class ActivationPlayableAsset extends PlayableAsset {
+  override createPlayable (): Playable {
+    return new ActivationPlayable();
+  }
+}
