+++ plugin-packages/model/src/plugin/model-item.ts
@@ -2,7 +2,7 @@ import type {
   Engine, HitTestBoxParams, HitTestCustomParams, HitTestSphereParams, Renderer, VFXItem,
 } from '@galacean/effects';
 import {
-  HitTestType, Behaviour, RendererComponent, effectsClass, spec, AnimationClip,
+  HitTestType, Behaviour, RendererComponent, effectsClass, spec,
 } from '@galacean/effects';
 import type {
   ModelCameraComponentData, ModelItemBounding, ModelLightComponentData,
@@ -15,6 +15,7 @@ import type { Euler, Ray, Vector2 } from '../runtime/math';
 import { Vector3 } from '../runtime/math';
 import { RayIntersectsBoxWithRotation } from '../utility';
 import { getSceneManager } from './model-plugin';
+import { AnimationClip } from 'packages/effects-core/src/animation';
 
 /**
  * 插件 Mesh 组件类，支持 3D Mesh 渲染能力
